actualizacion de precios con decimales

+++ src/main/java/com/camilo/fitnorius/dto/ProductDTO.java
@@ -1,19 +1,26 @@
 package com.camilo.fitnorius.dto;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ProductDTO {
+
     private Long id;
     private String name;
-    private Double price;
-    private Double oldPrice;
-    private Double discount;
+
+    // 💰 Usamos BigDecimal para precios exactos
+    private BigDecimal price;
+    private BigDecimal oldPrice;
+    private BigDecimal discount;
+
     private String imageUrl;
     private Long categoryId;
     private String categoryName;
-    private String description; // ✅ Nuevo campo para descripción del producto
+
+    // 📝 Descripción opcional del producto
+    private String description;
 }
